repurpose user.search controller method
