+++ server/routes/user.js
@@ -3,6 +3,6 @@ const router = express.Router();
 const protect = require("../middleware/auth");
 const { searchUsers } = require("../controllers/user");
 
-router.route("/").get(protect, searchUsers);
+router.route('').get(protect, searchUsers);
 
 module.exports = router;
